pyproject.toml
[tool.poetry]
name = "scraper_api"
version = "0.1.0"
description = ""
authors = ["Miguel"]

[tool.poetry.dependencies]
python = "^3.10"
Scrapy = "^2.6.2"
python-dotenv = "^0.20.0"
fastapi = "^0.79.1"
uvicorn = "^0.18.2"
crochet = "^2.0.0"
growattServer = "^1.2.2"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
